Use correct reference in man page links
After the '.so' one is supposed to use the directory name, like we have
correctly in all old man pages:
  mysql_client_test_embedded.1:.so man1/mysql_client_test.1
  mysql_embedded.1:.so man1/mysql.1
  mysqltest_embedded.1:.so man1/mysqltest.1

This change adds the 'man1/' component so the link has the correct format.
Actually using man links is a deprecated practice and using symlinks
would be better, but that can be fixed in a later commit.

From https://www.debian.org/doc/debian-policy/ch-docs.html#manual-pages:
> If one man page needs to be accessible via several names it is better
> to use a symbolic link than the .so feature

Detected via Lintian errors:
  E: mariadb-server-10.5: bad-so-link-within-manual-page usr/share/man/man1/mariadb-show.1.gz
  E: mariadb-client-10.5: bad-so-link-within-manual-page usr/share/man/man1/mariadb-binlog.1.gz
  E: mariadb-client-10.5: bad-so-link-within-manual-page usr/share/man/man1/mariadb-plugin.1.gz

Related to MDEV-21769.