MDEV-21723 merging galera.galera_slave_replay test from 10.3
This pull request merges mtr test galera.galera_slave_replay from 10.3
Test cluster topology is somewhat different in 10.4, and the test had to be adapted to 10.4 pratices.

The test execution showed one issue in async replication, where SQL slave thread aborted if event applying hit a conflict with galera cluster replication.
If the conflict was such, that async replication transaction had to be relayed in galera cluster, then SQL slave thread aborted. However, as there was need to replay, it actually means that the async replication transaction was accepted and slave thread could (and should) continue.
This was fixed by clearing the error code, if async replication transaction is marked to replay.