Specify a new client_plugindir path, to be used during testing
When testing installed packages, the client (mariadb or mysql) can have
plugins stored in a separate location compared to server. This change
ensures that the correct paths are used for the plugin-dir directive in
the my.cnf generated by mysql-test-run.

The directives in my.cnf will be replaced like so:

[client]
plugin-dir=$client_plugindir

[mysqld]
plugin-dir=$plugindir

where $<variable-name> corresponds to the variable name in mtr.