MDEV-19344 innodb.innodb-change-buffer-recovery fails
The test was incompatible with ./mtr --repeat=2 until
commit 2d6719d7ee92843d5b0b9a27c7deaff5cacd4745
fixed that.

It turns out that the failing assertion that we disabled in
commit 3db94d2403c1f2791bddd43656d7b0d6320b453d
is bogus and can fail when the change buffer is emptied
during the last batch of crash recovery. The reason for this
is the condition around the page_create_empty() call in
page_cur_delete_rec(). The condition was removed in MariaDB
Server 10.5 as part of MDEV-12353, in
commit 7ae21b18a6b73bbc3bf1ff448faf60c29ac1d386 and
commit f8a9f906679e1d1ab026c245f7d24c652050d8b3.

The bug that the assertion aimed to catch is MDEV-22497, which
was fixed in commit 26aab96ecfc9eca647ab9d5b4be3ba5823df4cde.