MDEV-22554: "mariabackup --prepare" exits with code 0 even though innodb error is logged
The fix is to set flag in ib::error::~error() and check it in
mariabackup.

ib::error::error() is replaced to ib::warn::warn() in
AIO::linux_create_io_ctx() because of two reasons:

1) if we leave it as is, then mariabackup MTR tests with --mem MTR option will
fail, :because Linux AIO can not be used on tmpfs,

2) when Linux AIO can not be initialized, InnoDB falls back to simulated
AIO, so such sutiation is not fatal error, it should be treated as warning.