MDEV-21976: Re-disable main.udf, reverts mistake in merge b30a01314217
This test was disabled in a135f0ab88d63b9a8976d6b3010f27766c38873d, but
accidentally reverted during the merge in
https://github.com/MariaDB/server/compare/0186b0a07760...b30a01314217

Because it was a 3-way merge, the removal is not directly visible.
It can be verified manually though:

  Line not changed:
  git diff a135f0ab88d6..0186b0a07760 mysql-test/unstable-tests

  Line removed:
  git diff a135f0ab88d6..b30a01314217 mysql-test/unstable-tests


This commit fixes the Travis-CI build and makes the main 10.5 git branch
green again on Travis-CI (and whatever else build environment that also
is failing due to this test).

Related: MDEV-22592