MDEV-22545: One more fix: main.perror-win
In 10.3 and earlier, some lines in the result file were
terminated by CR and LF, but in 10.4, all files are expected
to be terminated by LF only.

Also, my build of 10.4 apparently failed without me noticing
it, and the MDEV-22545 fix was not present. The %T should
not have been replaced with %s at all.