MDEV-21469 Pre-commit: make sure binlog hton is listed in the head of the user XA participants list.
This patch makes sure that at handling of the user XA
the binlog hton heads the list of the transaction branches
which was not the case for the multi-engine XA.

Also a recovered XA's completion (XA COMMIT or ROLLBACK through a
"foreign" connection) is made to invoke first binlog hton's
"complete"_by_xid before the general hton loop (and naturally skip the
binlog hton action in the loop).

The work is a prerequisite for MDEV-21469.