MDEV-22552: Remove mytop as deprecated
For historical reasons the sources have carried a copy of the mytop script.
It is not being maintained in MariaDB sources and it was not intended for
distribution, so clarify the situation by completely removing it and let
users install it from the origin themselves.

Most Linux distros such as Debian and Fedora ship mytop as a separate
package which users should use.