MDEV-22773 Assertion page_get_page_no... in btr_pcur_store_position()
btr_pcur_store_position(): Replace a too strict debug assertion.
It is possible to have a clustered index B-tree for a logically
empty table, which will consist of a node pointer from the root
page to a leaf page that contains the metadata record.

The too strict debug assertion was added in
commit 0e5a4ac2532c64a545796c787354dc41d61d0e62 (MDEV-15562).